+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.0.17]
+
+### Changed
+- ApplicationDone flag renamed to IsApplied
+
+### Fixed
+- Stretched widgets were not updated after resizing parent
+
 ## [0.0.16]
 
 ### Added